modules: move ctrl functions to ctrl folder
emulator: add missing include guards.
