+++ src/emulator/ctrl/CMakeLists.txt
@@ -1,7 +1,11 @@
 add_library(
 	ctrl
-	INTERFACE
+	STATIC
+	include/ctrl/definitions.h
+	include/ctrl/functions.h
+	include/ctrl/state.h
+	src/ctrl.cpp
 )
 
-target_include_directories(ctrl INTERFACE include)
-target_link_libraries(ctrl INTERFACE sdl2)
+target_include_directories(ctrl PUBLIC include)
+target_link_libraries(ctrl PUBLIC sdl2 vita-headers)
