+++ src/emulator/module/include/module/load_module.h
@@ -15,6 +15,8 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
+#pragma once
+
 #include <host/state.h>
 #include <kernel/thread/thread_functions.h>
